move ICorePlaybackObservation in src/playbackObserver
