+++ src/core/cmcd/cmcd_data_builder.ts
@@ -1,5 +1,4 @@
 import log from "../../log";
-import type { ICorePlaybackObservation } from "../../main_thread/init/utils/create_core_playback_observer";
 import type {
   IAdaptation,
   IManifest,
@@ -8,6 +7,7 @@ import type {
   ISegment,
 } from "../../manifest";
 import type { IReadOnlyPlaybackObserver } from "../../playback_observer";
+import type { ICorePlaybackObservation } from "../../playback_observer/types";
 import type { ICmcdOptions, ICmcdPayload, ITrackType } from "../../public_types";
 import createUuid from "../../utils/create_uuid";
 import isNullOrUndefined from "../../utils/is_null_or_undefined";
