+++ src/main_thread/init/multi_thread_content_initializer.ts
@@ -33,6 +33,7 @@ import type {
   IReadOnlyPlaybackObserver,
   IMediaElementPlaybackObserver,
 } from "../../playback_observer";
+import type { ICorePlaybackObservation } from "../../playback_observer/types";
 import type {
   ICmcdOptions,
   IInitialManifest,
@@ -57,7 +58,7 @@ import type { ITextDisplayer } from "../text_displayer";
 import sendMessage from "./send_message";
 import type { ITextDisplayerOptions } from "./types";
 import { ContentInitializer } from "./types";
-import type { ICorePlaybackObservation } from "./utils/create_core_playback_observer";
+
 import createCorePlaybackObserver from "./utils/create_core_playback_observer";
 import { resetMediaElement } from "./utils/create_media_source";
 import type { IInitialTimeOptions } from "./utils/get_initial_time";
