+++ src/main_thread/init/utils/create_core_playback_observer.ts
@@ -15,7 +15,6 @@
  */
 
 import getBufferedDataPerMediaBuffer from "../../../core/main/common/get_buffered_data_per_media_buffer";
-import type { IPausedPlaybackObservation } from "../../../core/types";
 import type { IManifestMetadata } from "../../../manifest";
 import { getMaximumSafePosition } from "../../../manifest";
 import type { IMediaSourceInterface } from "../../../mse";
@@ -23,12 +22,8 @@ import type {
   IPlaybackObservation,
   IReadOnlyPlaybackObserver,
   IMediaElementPlaybackObserver,
-  ObservationPosition,
-  IRebufferingStatus,
-  IFreezingStatus,
 } from "../../../playback_observer";
-import type { ITrackType } from "../../../public_types";
-import type { IRange } from "../../../utils/ranges";
+import type { ICorePlaybackObservation } from "../../../playback_observer/types";
 import type { IReadOnlySharedReference } from "../../../utils/reference";
 import SharedReference from "../../../utils/reference";
 import type { CancellationSignal } from "../../../utils/task_canceller";
@@ -53,35 +48,6 @@ export interface ICorePlaybackObserverArguments {
   textDisplayer: ITextDisplayer | null;
   /** Used abstraction for MSE API. */
   mediaSource: IMediaSourceInterface | null;
-}
-
-export interface ICorePlaybackObservation {
-  /**
-   * Information on whether the media element was paused at the time of the
-   * Observation.
-   */
-  paused: IPausedPlaybackObservation;
-  /**
-   * Information on the current media position in seconds at the time of the
-   * Observation.
-   */
-  position: ObservationPosition;
-  /** `duration` property of the HTMLMediaElement. */
-  duration: number;
-  /** `readyState` property of the HTMLMediaElement. */
-  readyState: number;
-  /** Target playback rate at which we want to play the content. */
-  speed: number;
-  /** Theoretical maximum position on the content that can currently be played. */
-  maximumPosition: number;
-  /**
-   * Ranges of buffered data per type of media.
-   * `null` if no buffer exists for that type of media.
-   */
-  buffered: Record<ITrackType, IRange[] | null>;
-  rebuffering: IRebufferingStatus | null;
-  freezing: IFreezingStatus | null;
-  bufferGap: number | undefined;
 }
 
 /**
