+++ src/playback_observer/types.ts
@@ -1,3 +1,5 @@
+import type { ITrackType } from "../public_types";
+import type { IRange } from "../utils/ranges";
 import type { IReadOnlySharedReference } from "../utils/reference";
 import type { CancellationSignal } from "../utils/task_canceller";
 import type ObservationPosition from "./utils/observation_position";
@@ -235,4 +237,53 @@ export interface IReadOnlyPlaybackObserver<TObservationType> {
       cancellationSignal: CancellationSignal,
     ) => IReadOnlySharedReference<TDest>,
   ): IReadOnlyPlaybackObserver<TDest>;
+}
+
+export interface ICorePlaybackObservation {
+  /**
+   * Information on whether the media element was paused at the time of the
+   * Observation.
+   */
+  paused: IPausedPlaybackObservation;
+  /**
+   * Information on the current media position in seconds at the time of the
+   * Observation.
+   */
+  position: ObservationPosition;
+  /** `duration` property of the HTMLMediaElement. */
+  duration: number;
+  /** `readyState` property of the HTMLMediaElement. */
+  readyState: number;
+  /** Target playback rate at which we want to play the content. */
+  speed: number;
+  /** Theoretical maximum position on the content that can currently be played. */
+  maximumPosition: number;
+  /**
+   * Ranges of buffered data per type of media.
+   * `null` if no buffer exists for that type of media.
+   */
+  buffered: Record<ITrackType, IRange[] | null>;
+  rebuffering: IRebufferingStatus | null;
+  freezing: IFreezingStatus | null;
+  bufferGap: number | undefined;
+}
+
+/** Pause-related information linked to an emitted Playback observation. */
+export interface IPausedPlaybackObservation {
+  /**
+   * Known paused state at the time the Observation was emitted.
+   *
+   * `true` indicating that the HTMLMediaElement was in a paused state.
+   *
+   * Note that it might have changed since. If you want truly precize
+   * information, you should recuperate it from the HTMLMediaElement directly
+   * through another mean.
+   */
+  last: boolean;
+  /**
+   * Actually wanted paused state not yet reached.
+   * This might for example be set to `false` when the content is currently
+   * loading (and thus paused) but with autoPlay enabled.
+   */
+  pending: boolean | undefined;
 }
